added IPaginationOptions abstraction
