+++ src/Modules/InventoryManagement/DigiDent.InventoryManagement.Domain/Items/ReadModels/InventoryItemDetails.cs
@@ -4,7 +4,7 @@ using DigiDent.InventoryManagement.Domain.Requests.ReadModels;
 namespace DigiDent.InventoryManagement.Domain.Items.ReadModels;
 
 public record InventoryItemDetails(
-    Guid Id,
+    int Id,
     string Name,
     string Remarks,
     int Quantity,
@@ -12,3 +12,6 @@ public record InventoryItemDetails(
     List<RequestSummary> OpenRequests,
     List<ActionSummary> LastActions
 );
+//TODO: probably move to the application layer,
+//since the mapping will be from the Domain Entity
+//returned form GetById repository method
