+++ src/Shared/DigiDent.Shared.Infrastructure/Pagination/PaginationDTO.cs
@@ -1,0 +1,10 @@
+﻿using DigiDent.Shared.Kernel.ValueObjects.Pagination;
+
+namespace DigiDent.Shared.Infrastructure.Pagination;
+
+public record PaginationDTO(
+    int PageNumber = 1,
+    int PageSize = 20,
+    string SearchTerm = "",
+    string SortByColumn = "Id",
+    SortOrder SortOrder = SortOrder.Desc): IPaginationOptions;
